# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Farmers Market Platform API",
    "version": "1.0.0",
    "description": "Divine Agricultural Platform - Comprehensive REST API\n\nThis API provides endpoints for:\n- 🚜 Farm Management (CRUD, search, verification)\n- 🌾 Product Catalog (inventory, categories, search)\n- 📦 Order Processing (cart, checkout, fulfillment)\n- 👤 User Management (authentication, profiles, roles)\n- 🌍 Marketplace (discovery, featured items, recommendations)\n- 📊 Analytics (farmer dashboard, sales reports)\n- 🔔 Notifications (real-time updates, alerts)\n\nAll endpoints return standardized ServiceResponse<T> format with:\n- success: boolean\n- data?: T\n- error?: ServiceError\n- meta?: ResponseMetadata (including agricultural consciousness)",
    "contact": {
      "name": "API Support",
      "email": "support@farmersmarket.com",
      "url": "https://farmersmarket.com/support"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/api/farms/{id}": {
      "get": {
        "tags": ["FarmController"],
        "summary": "handleGetFarm",
        "description": "Get farm by ID",
        "operationId": "get__api_farms_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Farm ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Farm details"
          },
          "404": {
            "description": "Farm not found"
          }
        }
      },
      "put": {
        "tags": ["FarmController"],
        "summary": "handleUpdateFarm",
        "description": "Update farm (requires ownership)",
        "operationId": "put__api_farms_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Farm ID"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Farm updated successfully"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not the farm owner"
          }
        }
      },
      "delete": {
        "tags": ["FarmController"],
        "summary": "handleDeleteFarm",
        "description": "Delete farm (soft delete, requires ownership)",
        "operationId": "delete__api_farms_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Farm ID"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Farm deleted successfully"
          }
        }
      }
    },
...
    "/api/products/search": {
      "get": {
        "tags": ["ProductController"],
        "summary": "handleSearchProducts",
        "description": "Search products by keyword",
        "operationId": "get__api_products_search",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Search query"
          }
        ],
        "responses": {
          "200": {
            "description": "Search results"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from NextAuth"
      }
    },
...
  }
}